Move CPlayerSkills traits
