+++ include/football/CPlayerSkills.h
@@ -1,32 +1,11 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CPlayerSkills.h"
 
 #include "football/types/CPlayerSkills.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CPlayerSkills;
-}
-
-template <> struct json_traits<football::CPlayerSkills>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Player skills";
-	//! JSON key for a skill.
-	template <football::E_PLAYER_SKILL tPlayerSkill> static inline constexpr std::string_view SKILL_KEY{};
-	//! JSON key for a skill experience.
-	template <football::E_PLAYER_SKILL tPlayerSkill> static inline constexpr std::string_view XP_KEY{};
-};
-
-FOR_EACH_PLAYER_SKILL( FUTSIM_INSTANTIATE_ENUM_TRAIT, football::CPlayerSkills, SKILL_KEY, football::E_PLAYER_SKILL, " skill" )
-FOR_EACH_PLAYER_SKILL( FUTSIM_INSTANTIATE_ENUM_TRAIT, football::CPlayerSkills, XP_KEY, football::E_PLAYER_SKILL, " experience" )
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -83,6 +62,4 @@ private:
 	experiences mExperiences;
 };
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
